Quantities can have decimals

+++ lib/qrda-import/base-importers/section_importer.rb
@@ -125,8 +125,8 @@ module QRDA
         return unless value_element && !value_element['nullFlavor']
         value = value_element['value']
         if value.present?
-          return value.strip.to_i if (value_element['unit'] == "1" || value_element['unit'].nil?)
-          return QDM::Quantity.new(value.strip.to_i, value_element['unit'])
+          return value.strip.to_f if (value_element['unit'] == "1" || value_element['unit'].nil?)
+          return QDM::Quantity.new(value.strip.to_f, value_element['unit'])
         elsif value_element['code'].present?
           return code_if_present(value_element)
         end
@@ -160,7 +160,7 @@ module QRDA
       def extract_scalar(parent_element, scalar_xpath)
         scalar_element = parent_element.at_xpath(scalar_xpath)
         return unless scalar_element
-        QDM::Quantity.new(scalar_element['value'].to_i, scalar_element['unit'])
+        QDM::Quantity.new(scalar_element['value'].to_f, scalar_element['unit'])
       end
 
       def extract_components(parent_element)
